Fix: timeSpent double (ce9bf91fdc6c051ad2ab029535f3b4468c7af3c2)
Forgot to add in ce9bf91fdc6c051ad2ab029535f3b4468c7af3c2

+++ tests/suite/Manager.cxx
@@ -89,7 +89,7 @@ namespace suite {
     // (as it was recorded).
     writeUpdateStats data {
       .timeRequired = frameTime,
-      .timeSpent = (uint) (time.count() * 10e2),
+      .timeSpent = time.count() * 10e2,
     };
     timedEncoder_->addWriteUpdate(data);
   }
